requirements.txt
langchain
langchain-community
langchain-chroma
langchain-openai
chromadb
tiktoken
pypdf
python-dotenv
streamlit
openai